+++ university-management-admin-api/src/main/java/com/wubo/api/entity/Admission.java
@@ -13,9 +13,9 @@ import lombok.NoArgsConstructor;
 @TableName("admission") // 指定数据库表名
 public class Admission {
     @TableId(value = "id", type = IdType.AUTO) // 主键，自增
-    private Long id;
+    private Integer id;
 
-    private Long universityId; // 外键，指向高校表
+    private Integer universityId; // 外键，指向高校表
     private String province; // 省份
     private int year; // 年份
     private int scoreRequired; // 录取分数线
