+++ university-management-admin-api/src/main/java/com/wubo/api/entity/Feedback.java
@@ -13,9 +13,9 @@ import lombok.NoArgsConstructor;
 @TableName("feedback") // 指定数据库表名
 public class Feedback {
     @TableId(value = "id", type = IdType.AUTO) // 主键，自增
-    private Long id;
+    private Integer id;
 
-    private Long universityId; // 外键，指向高校表
-    private Long userId; // 外键，指向用户表
+    private Integer universityId; // 外键，指向高校表
+    private Integer userId; // 外键，指向用户表
     private String feedbackText; // 用户反馈内容
 }
