+++ university-management-admin-api/src/main/java/com/wubo/api/entity/RecommendationRating.java
@@ -13,9 +13,9 @@ import lombok.NoArgsConstructor;
 @TableName("recommendation_rating") // 指定数据库表名
 public class RecommendationRating {
     @TableId(value = "id", type = IdType.AUTO) // 主键，自增
-    private Long id;
+    private Integer id;
 
-    private Long universityId; // 外键，指向高校表
+    private Integer universityId; // 外键，指向高校表
     private String majorCategory; // 专业类别
     private double rating; // 推荐指数
     private int recommendationCount; // 推荐人数
